Переносит жанр и активный фильм из store.DATA в store.APP_STATE
