+++ src/reducer/app-state/app-state.js
@@ -1,13 +1,18 @@
 import {extend} from "../../utils.js";
-import {SHOWED_MOVIES_ON_START_COUNT} from "../../const.js";
+import {DEFAULT_GENRE, SHOWED_MOVIES_ON_START_COUNT} from "../../const.js";
 
 const initialState = {
+  activeMovie: {},
+  genre: DEFAULT_GENRE,
   moviesCount: SHOWED_MOVIES_ON_START_COUNT,
 };
 
 const ActionType = {
   INCREMENT_MOVIES_COUNT: `INCREMENT_MOVIES_COUNT`,
+  RESET_ACTIVE_MOVIE: `RESET_ACTIVE_MOVIE`,
   RESET_COUNT: `RESET_COUNT`,
+  SET_ACTIVE_MOVIE: `SET_ACTIVE_MOVIE`,
+  SET_GENRE: `SET_GENRE`,
 };
 
 const ActionCreator = {
@@ -16,10 +21,24 @@ const ActionCreator = {
     payload: SHOWED_MOVIES_ON_START_COUNT
   }),
 
+  resetActiveMovie: () => ({
+    type: ActionType.RESET_ACTIVE_MOVIE,
+  }),
+
   resetCount: () => ({
     type: ActionType.RESET_COUNT,
     payload: null,
-  })
+  }),
+
+  setActiveMovie: (movie) => ({
+    type: ActionType.SET_ACTIVE_MOVIE,
+    payload: movie,
+  }),
+
+  setGenre: (genre) => ({
+    type: ActionType.SET_GENRE,
+    payload: genre,
+  }),
 };
 
 const reducer = (state = initialState, action) => {
@@ -31,6 +50,18 @@ const reducer = (state = initialState, action) => {
     case ActionType.RESET_COUNT:
       return extend(state, {
         moviesCount: SHOWED_MOVIES_ON_START_COUNT,
+      });
+    case ActionType.RESET_ACTIVE_MOVIE:
+      return extend(state, {
+        activeMovie: state.promoMovie.id,
+      });
+    case ActionType.SET_ACTIVE_MOVIE:
+      return extend(state, {
+        activeMovie: action.payload,
+      });
+    case ActionType.SET_GENRE:
+      return extend(state, {
+        genre: action.payload,
       });
   }
   return state;
