+++ src/reducer/app-state/selectors.js
@@ -6,3 +6,12 @@ export const getMoviesCount = (state) => {
   return state[NAME_SPACE].moviesCount;
 };
 
+export const getActiveMovie = (state) => {
+  return state[NAME_SPACE].activeMovie;
+};
+
+export const getGenre = (state) => {
+  return state[NAME_SPACE].genre;
+};
+
+
